+++ cli/main.go
@@ -30,10 +30,14 @@ func main() {
 		*CreatePipelineTemplateCommand(),
 		*UpdateAgentCommand(),
 		*UpdateAgentsCommand(),
+		*UpdatePipelineConfigCommand(),
 		*UpdatePipelineTemplateCommand(),
 		*DeleteAgentCommand(),
 		*DeleteAgentsCommand(),
 		*DeletePipelineTemplateCommand(),
+		*ListPipelineGroupsCommand(),
+		*GetPipelineHistoryCommand(),
+		*CreatePipelineConfigCommand(),
 	}
 
 	app.Flags = []cli.Flag{