+++ gocd.go
@@ -48,6 +48,8 @@ type Client struct {
 	Stages            *StagesService
 	Jobs              *JobsService
 	PipelineTemplates *PipelineTemplatesService
+	Pipelines         *PipelinesService
+	PipelineConfigs   *PipelineConfigsService
 
 	common service
 	cookie string
@@ -92,6 +94,8 @@ func NewClient(gocdBaseUrl string, auth *Auth, httpClient *http.Client, checkSsl
 	c.Stages = (*StagesService)(&c.common)
 	c.Jobs = (*JobsService)(&c.common)
 	c.PipelineTemplates = (*PipelineTemplatesService)(&c.common)
+	c.Pipelines = (*PipelinesService)(&c.common)
+	c.PipelineConfigs = (*PipelineConfigsService)(&c.common)
 	return c
 }
 
@@ -102,12 +106,7 @@ func (c *Client) NewRequest(method, urlStr string, body interface{}, apiVersion 
 		return nil, err
 	}
 
-	if apiVersion == "" {
-		apiVersion = apiV1
-	}
-
 	u := c.BaseURL.ResolveReference(rel)
-
 	request := &APIRequest{}
 
 	var buf io.ReadWriter
@@ -136,7 +135,9 @@ func (c *Client) NewRequest(method, urlStr string, body interface{}, apiVersion 
 	if body != nil {
 		req.Header.Set("Content-Type", "application/json")
 	}
-	req.Header.Set("Accept", apiVersion)
+	if apiVersion != "" {
+		req.Header.Set("Accept", apiVersion)
+	}
 	req.Header.Set("User-Agent", c.UserAgent)
 
 	if c.cookie == "" {